fix(fe): 로그인 페이지에서 페이지 이동 시 발생하는 이벤트 등록
로그인 폼은 페이지 이동 시 발생하는 이벤트 등록을 제거하여 폼의 역할만 하도록 함
