+++ web/components/Forms/LoginForm/index.js
@@ -1,4 +1,4 @@
-import React, { useContext, useEffect } from 'react';
+import React from 'react';
 import Router from 'next/router';
 import useForm from 'react-hook-form';
 
@@ -12,26 +12,9 @@ import validator from '../../../utils/validator';
 import S from './styled';
 import storage from '../../../utils/storage';
 import request from '../../../utils/request';
-import { setMessage } from '../../../contexts/reducer';
-import { AppDispatchContext } from '../../../contexts';
-
-const subscribeRouteChangeStartFunction = dispatch => {
-  const handleRouteChange = url => {
-    if (url !== '/login') {
-      dispatch(setMessage(''));
-      Router.events.off('routeChangeStart', handleRouteChange);
-    }
-  };
-  Router.events.on('routeChangeStart', handleRouteChange);
-};
 
 const LoignForm = () => {
   const { register, handleSubmit, errors, setError, clearError } = useForm();
-  const { dispatch } = useContext(AppDispatchContext);
-
-  useEffect(() => {
-    subscribeRouteChangeStartFunction(dispatch);
-  }, [dispatch]);
 
   const onSubmit = (data, e) => {
     const { userId, password } = data;
